% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_predictor_binomial.R
\name{best_predictor_binomial}
\alias{best_predictor_binomial}
\title{Using bootstrapping and LASSO algorithm to choose best predictive features}
\usage{
best_predictor_binomial(
  target_data,
  response = "response",
  features,
  target_data_id = "ID",
  features_id = "ID",
  show_progress = TRUE,
  permutation = 1000,
  propotion = 0.8,
  nfolds = 10,
  plot_vars = 20,
  color = "#999999",
  palette = "Blues",
  discrete_x = 20
)
}
\arguments{
\item{target_data}{Data frame contains patient identifier, and target_data}

\item{response}{binary variables}

\item{features}{Data frame contains patient identifier and variables after feature engineering}

\item{target_data_id}{identifier of phenotype data}

\item{features_id}{identifier of feature matrix}

\item{show_progress}{show progress bar}

\item{permutation}{times of iteration}

\item{propotion}{proportion of patients in each bootstrapping iteration}

\item{nfolds}{folds to perform cross validation in LASSO}

\item{plot_vars}{plotting important variables}

\item{color}{default is steelblue}

\item{palette}{plotting palette, default is `#999999`, using `RColorBrewer::display.brewer.all()` to see more options}

\item{discrete_x}{if maximal character length of variables is larger than discrete_x, label will be discrete}
}
\value{

}
\description{
Using bootstrapping and LASSO algorithm to choose best predictive features
}
\examples{
res<-best_predictor_binomial(target_data = target, features = features,response = "status",nfolds = 10,permutation = 100)
}
\author{
Dongqiang Zeng
}
