% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLasso_Gibbs_Joo.R
\name{BLasso}
\alias{BLasso}
\title{Bayesian Lasso by Gibbs Sampler}
\usage{
BLasso(x, y, n.max = 10000, EB = TRUE, a = 1, b = 1, print.it = FALSE)
}
\arguments{
\item{a, }{b: specify these for a hyper-Gamma prior for lambda^2 if EB = FALSE}

\item{print.it}{= TRUE/FALSE (default: FALSE, suppressing to print the number of iterations)}

\item{x:}{predictor variables (numertic only)}

\item{y:}{outcome (numertic only)}

\item{n.max:}{n of interations (default: 10000)}

\item{EB:}{TRUE/FALSE (default: TRUE, estimating lambda by empircal bayes)}
}
\value{
beta

beta.95q: 95 \% (posterior) CI of beta

beta.sig: standard deviation of beta

tau2.95q: 95 \% CI of tau2

sigma.95q: 95 \% CI of sigma2

lambda: penalty, or global scale

lambda.95q: 95% CI of lambda
}
\description{
Provide two options for the estimation of penalty parameter "lambda"
}
\examples{
ex1<-BLasso(x=data1[,-1], y=data1[,1]); ex1$beta; ex1$lambda; sum(ex1$beta.95q[1,]*ex1$beta.95q[2,]>0); #n of selected variables#
}
\references{
Park, Trevor, and George Casella. "The bayesian lasso." Journal of the American Statistical Association 103.482 (2008): 681-686.
}

