% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{data1}
\alias{data1}
\title{data1 - diabetes \{lars\}}
\format{A data frame with 442 rows and 11 variables:

\describe{
  \item{y}{}
  \item{age}{}
  \item{sex}
  \item{bmi}
  \item{map}
  \item{tc}
  \item{ldl}
  \item{hdl}
  \item{tch}
  \item{ltg}
  \item{glu}
}}
\source{
\{lars\}
}
\usage{
data1
}
\description{
A dataset containing y and 10 predictors of 442 patients.
}
\keyword{datasets}

