% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{data2}
\alias{data2}
\title{data2 - prostate \{ElemStatLearn\}}
\format{A data frame with 97 rows and 9 variables:

\describe{
  \item{y}{}
  \item{lcavol}{}
  \item{lweight}{}
  \item{age}{}
  \item{lbph}{}
  \item{svi}{}
  \item{gleason}{}
  \item{pgg45}{}
}}
\source{
\{ElemStatLearn\}
}
\usage{
data2
}
\description{
A dataset containing y and 10 predictors of 97 cancer patients.
}
\keyword{datasets}

