% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data3.R
\docType{data}
\name{data3}
\alias{data3}
\title{data3 - ozeon \{mbelch\}}
\format{A data frame with 303 rows and 14 variables:

\describe{
  \item{y}{}
  \item{Month}{}
  \item{Day.m}{}
  \item{Day.w}{}
  \item{Pressure.Height}{}
  \item{Wind.Speed}{}
  \item{Humidity}{}
  \item{Temp.Sandburg.}{}
  \item{Temp.ElMonte.}{}
  \item{Inversion.height}{}
  \item{Pressure.Gradient}{}
  \item{Visibility}{}
}}
\source{
\{mbelch\}
}
\usage{
data3
}
\description{
A dataset containing y and 13 predictors of 203 measurements.
}
\keyword{datasets}

