% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data5.R
\docType{data}
\name{data5}
\alias{data5}
\title{data5 - Wine Quality Data - UCI - Machine Learning Directory}
\format{A data frame with 6497 rows and 13 variables:

\describe{
  \item{y}{wine quality score 1-10}
  \item{type}{1: red, 0: white}
  \item{fixed.acidity}{}
  \item{volatile.acidity}{}
  \item{citric.acid}{}
  \item{residual.sugar}{}
  \item{chlorides}{}
  \item{free.sulfur.dioxide}{}
  \item{total.sulfur.dioxide}{}
  \item{density}{}
  \item{pH}{}
  \item{sulphates}{}
  \item{alcohol}{}
}}
\source{
\url<https://archive.ics.uci.edu/ml/datasets/Wine+Quality>
}
\usage{
data5
}
\description{
A dataset containing 6497 wine scores with 12 wine attributes variables.
}
\keyword{datasets}

