% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLasso_DL_Joo.R
\name{DLasso}
\alias{DLasso}
\title{Bayesian Lasso by Variational Bayes + Dirichlet-Laplace Priors}
\usage{
DLasso(x, y, print.it = FALSE)
}
\arguments{
\item{print.it}{= TRUE/FALSE (default: FALSE, suppressing to print the number of iterations)}

\item{x:}{predictor variables (numertic only)}

\item{y:}{outcome (numertic only)}
}
\value{
beta

beta.sig: standard deviation of beta

p.value: p-value for t-test for beta (for variable selection)

sigma2

lambda: penalty, or global scale

convergence: 1/0 converged if convergence = 1
}
\description{
Extended Bayesian Lasso (Park \& Casella (2008)) with Dirichlet-Laplace Priors (Bhattacharya et al. 2015)
}
\examples{
ex1<-DLasso(x=data1[,-1], y=data1[,1]); ex1$beta; ex1$lambda; sum(ex1$p.value<0.05); #n of selected variables#

}
\references{
Bhattacharya, Anirban, et al. "Dirichlet–Laplace priors for optimal shrinkage." Journal of the American Statistical Association 110.512 (2015): 1479-1490.

Park, Trevor, and George Casella. "The bayesian lasso." Journal of the American Statistical Association 103.482 (2008): 681-686.

Joo, Lijin. "Bayesian Lasso: An Extension for Genome-wide Assoication Study." New York University, 2017
}

