% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_interactions.R
\name{find_interaction}
\alias{find_interaction}
\title{Find interactions}
\usage{
find_interaction(input, num_target = 2)
}
\arguments{
\item{input}{list of variables chosen by the model}

\item{num_target}{default is 2}
}
\description{
Find interactions
}
\examples{
res<-best_predictor_binomial(target_data = target, features = features,response = "status",nfolds = 10,permutation = 100)
inter<-find_interaction(input = res$res_list, num_target = 2)
}
