% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLasso_VB_Joo.R
\name{VBLasso}
\alias{VBLasso}
\title{Bayesian Lasso by Variational Bayes}
\usage{
VBLasso(x, y, print.it = FALSE)
}
\arguments{
\item{print.it}{= TRUE/FALSE (default: FALSE, suppressing to print the number of iterations)}

\item{x:}{predictor variables (numertic only)}

\item{y:}{outcome (numertic only)}
}
\value{
beta

beta.sig: standard deviation of beta

p.value: p-value for t-test for beta (for variable selection)

sigma2

lambda: penalty, or global scale

convergence: 1/0 converged if convergence = 1
}
\description{
Fit Bayesian Lasso (Park \& Casella (2008)) by variational Bayes
}
\examples{
ex1<-VBLasso(x=data1[,-1], y=data1[,1]); ex1$beta; ex1$lambda; sum(ex1$p.value<0.05); #n of selected variables#

}
\references{
Park, Trevor, and George Casella. "The bayesian lasso." Journal of the American Statistical Association 103.482 (2008): 681-686.

Joo, Lijin. "Bayesian Lasso: An Extension for Genome-wide Assoication Study." New York University, 2017
}

