% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_predictor_binomial.R
\name{best_predictor_binomial}
\alias{best_predictor_binomial}
\title{Using bootstraping and LASSO algorithm to choose best predictive features}
\usage{
best_predictor_binomial(
  target_data,
  response = "response",
  features,
  target_data_id = "ID",
  features_id = "ID",
  permutation = 1000,
  propotion = 0.8,
  nfolds = 10,
  plot_vars = 20,
  palette = "Blues"
)
}
\arguments{
\item{target_data}{Data frame contains patient identifer, and target_data}

\item{response}{binary variables}

\item{features}{Data frame contains patient identifer and variables after feature engineering}

\item{target_data_id}{column name of phenotype data}

\item{features_id}{column name of feature matrix}

\item{permutation}{times of iteration}

\item{propotion}{propotion of patients in each boostraping iteration}

\item{nfolds}{folds to perform cross validation in LASSO}

\item{plot_vars}{plotting important variables}

\item{palette}{plotting palette, using `RColorBrewer::brewer.pal()`}
}
\value{

}
\description{
Using bootstraping and LASSO algorithm to choose best predictive features
}
\examples{
res<-best_predictor_binomial(target_data = target, features = features,response = "status",nfolds = 10,permutation = 100)
}
\author{
Dongqiang Zeng
}
