% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_predictot_cox.R
\name{best_predictor_cox}
\alias{best_predictor_cox}
\title{Using bootstrapping and LASSO algorithm to choose best prognostic features}
\usage{
best_predictor_cox(
  target_data,
  features,
  status,
  time,
  target_data_id = "ID",
  features_id = "ID",
  permutation = 1000,
  propotion = 0.8,
  nfolds = 10,
  plot_vars = 20,
  palette = "Blues",
  show_progress = TRUE,
  discrete_x = 20
)
}
\arguments{
\item{target_data}{Data frame contains patient identifier, survival time and survival event = 0/1}

\item{features}{Data frame contains patient identifier and variables after feature engineering}

\item{status}{column name of survival event}

\item{time}{column name of survival time}

\item{target_data_id}{column name of phenotype data}

\item{features_id}{column name of feature matrix}

\item{permutation}{times of iteration}

\item{propotion}{proportion of patients in each bootstrapping iteration}

\item{nfolds}{folds to perform cross validation in LASSO}

\item{palette}{plotting palette, using `RColorBrewer::brewer.pal()`}

\item{show_progress}{show progress bar}

\item{discrete_x}{if maximal character length of variables is larger than discrete_x, label will be discrete}
}
\value{
variables with frequency
}
\description{
Using bootstrapping and LASSO algorithm to choose best prognostic features
}
\examples{
data("target")
data("features")
}
\author{
Dongqiang Zeng
}
